% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21017R.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and Cpp functions (\code{gibbsC}.
}
\examples{
\dontrun{
tm1 <- microbenchmark::microbenchmark(
  rnR = gibbsR(2000,1,4,50,10),
  rnC = gibbsC(2000,1,4,50,10)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
