% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21017R.R
\name{sample_tetraploid}
\alias{sample_tetraploid}
\title{Generate nucleotide read counts for tetraploids in SeqEM.}
\usage{
sample_tetraploid(n, lambda, alpha, p)
}
\arguments{
\item{n}{sample size}

\item{lambda}{assume read depth distributed as Poisson with mean lambda}

\item{alpha}{nucleotide-read error}

\item{p}{minor allele frequency}
}
\value{
a list for reads of size \code{n}
}
\description{
Generate read depths and variant counts with a symmetric error assumption.
}
\examples{
\dontrun{
sample_tetraploid(10,10,0.01,0.1)
}
}
