% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21017R.R
\name{seqem_tetraploid}
\alias{seqem_tetraploid}
\title{SeqEM method to estimate parameters for tetraploid individuals}
\usage{
seqem_tetraploid(reads, iters, epsilon)
}
\arguments{
\item{reads}{a data set for read sequences}

\item{iters}{maximum iterations}

\item{epsilon}{convergence criterion}
}
\value{
estimation of \code{alpha} and \code{p}
}
\description{
EM algorithm is used to estimate nucleotide-read error and MAF
}
\examples{
\dontrun{
sample <- sample_tetraploid(10,10,0.01,0.1)
seqem_tetraploid(sample, 100, 1e-8)
}
}
