\name{StatComp21017-package}
\alias{StatComp21017-package}
\alias{StatComp21017}
\docType{package}
\title{
  Project for 'statistical computing' course by Hanyun Liu
}
\description{
  R and Rcpp versions of gibbs sampler are provided in this package (\code{gibbsR} versus \code{gibbsC})
}
\author{
Hanyun Liu

Maintainer: Hanyun Liu <2796427840@qq.com>
}
\keyword{ package }