% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21017R.R
\name{trials_diploid}
\alias{trials_diploid}
\title{Simulations to assess the accuracy of SeqEM}
\usage{
trials_diploid(m, n, lambda, alpha, p)
}
\arguments{
\item{m}{repeat times}

\item{n}{sample size}

\item{lambda}{assume read depth distributed as Poisson with mean lambda}

\item{alpha}{nucleotide-read error}

\item{p}{minor allele frequency}
}
\value{
A matrix of estimates
}
\description{
Parameters are estimated using SeqEM
}
\examples{
\dontrun{
tr <- trials_diploid(1000,50,10,0.01,0.2)
apply(tr,1,mean)
apply(tr,1,sd)
}
}
